+++ csharp/Euler/p0011.cs
@@ -31,10 +31,10 @@ namespace Euler
                 for (byte j = 0; j < 17; j++)
                 {
                     // horizontal section
-                    tmp = grid[i][j] * grid[i][j + 1] * grid[i][j + 2] * grid[i][j + 3];
+                    tmp = (uint)grid[i][j] * grid[i][j + 1] * grid[i][j + 2] * grid[i][j + 3];
                     answer = Math.Max(answer, tmp);
                     // vertical section
-                    tmp = grid[j][i] * grid[j + 1][i] * grid[j + 2][i] * grid[j + 3][i];
+                    tmp = (uint)grid[j][i] * grid[j + 1][i] * grid[j + 2][i] * grid[j + 3][i];
                     answer = Math.Max(answer, tmp);
                 }
             }
@@ -43,10 +43,10 @@ namespace Euler
                 for (byte j = 0; j < 17; j++)
                 {
                     // right diagonal section
-                    tmp = grid[i][j] * grid[i + 1][j + 1] * grid[i + 2][j + 2] * grid[i + 3][j + 3];
+                    tmp = (uint)grid[i][j] * grid[i + 1][j + 1] * grid[i + 2][j + 2] * grid[i + 3][j + 3];
                     answer = Math.Max(answer, tmp);
                     // left diagonal section
-                    tmp = grid[i][j + 3] * grid[i + 1][j + 2] * grid[i + 2][j + 1] * grid[i + 3][j];
+                    tmp = (uint)grid[i][j + 3] * grid[i + 1][j + 2] * grid[i + 2][j + 1] * grid[i + 3][j];
                     answer = Math.Max(answer, tmp);
                 }
             }
